+++ CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `observation.id_atlas_code` returns the ornitho id prefixed with `3_`
 - `sites` can now be retrieved directly via api endpoint 
 - raise `ObjectNotFoundException` instead of `APIException` if no object is retrieved
+- `observation.diff` returns a dictionary containing the updated and deleted observations
 
 ### Fixed